exclusao de arquivos de produtos que nao é mais utilizado, organizacao de pastas para o componente pages onde se localiza o arquivo de produtos, exclusao de mais comentarios, criacao da pasta pages e styled para melhor organizacao dos arquivos
