+++ src/app/layout.js
@@ -13,8 +13,8 @@ const geistMono = localFont({
 });
 
 export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "My Side",
+  description: "My Side - Your next favorite side project",
 };
 
 export default function RootLayout({ children }) {
