--- src/app/product.module.css
@@ -1,9 +1,0 @@
-.container {
-    padding: 20px;
-  }
-  
-  .image {
-    max-width: 100%;
-    height: auto;
-  }
-
